- update ck release number - update tapestry dependencies (5.3.6) - move image service into core moule - remove ck-image-module
