+++ chenillekit-demo/pom.xml
@@ -26,7 +26,7 @@
 	<parent>
 		<groupId>org.chenillekit</groupId>
 		<artifactId>chenillekit-project</artifactId>
-		<version>1.3.4-alpha-1</version>
+		<version>1.3.4-alpha-2</version>
 	</parent>
 
 	<name>Chenillekit Tapestry Demo</name>
@@ -72,12 +72,6 @@
 		<dependency>
 			<groupId>${project.groupId}</groupId>
 			<artifactId>chenillekit-tapestry</artifactId>
-			<version>${project.parent.version}</version>
-		</dependency>
-
-		<dependency>
-			<groupId>${project.groupId}</groupId>
-			<artifactId>chenillekit-image</artifactId>
 			<version>${project.parent.version}</version>
 		</dependency>
 
@@ -133,12 +127,6 @@
 			<artifactId>servlet-api</artifactId>
 			<version>2.5</version>
 			<scope>provided</scope>
-		</dependency>
-
-		<dependency>
-			<groupId>com.google.code.kaptcha</groupId>
-			<artifactId>kaptcha</artifactId>
-			<version>2.3</version>
 		</dependency>
 
 	</dependencies>
