+++ chenillekit-demo/src/main/webapp/WEB-INF/conf/menu.demo.conf
@@ -22,6 +22,5 @@ Thumbnail       |tapcomp/ThumbnailDemo          |text-align:center;width:100px
 Tooltip         |tapcomp/TooltipDemo            |text-align:center;width:100px
 Slider          |tapcomp/SliderDemo             |text-align:center;width:100px
 Window          |tapcomp/WindowDemo             |text-align:center;width:100px
-Kaptcha         |tapcomp/KaptchaDemo            |text-align:center;width:100px
 Yahoo-Comps     |tapcomp/YuiComponentsDemo      |text-align:center;width:100px
 some other      |tapcomp/MultiCompDemo          |text-align:center;width:100px
