+++ chenillekit-google/pom.xml
@@ -26,7 +26,7 @@
     <parent>
         <groupId>org.chenillekit</groupId>
         <artifactId>chenillekit-project</artifactId>
-        <version>1.3.4-alpha-1</version>
+        <version>1.3.4-alpha-2</version>
     </parent>
 
     <name>Chenillekit Google</name>
