--- chenillekit-image/pom.xml
@@ -1,114 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-
-<!--
-  ~ Apache License
-  ~ Version 2.0, January 2004
-  ~ http://www.apache.org/licenses/
-  ~
-  ~ Copyright 2008-2012 by chenillekit.org
-  ~
-  ~ Licensed under the Apache License, Version 2.0 (the "License");
-  ~ you may not use this file except in compliance with the License.
-  ~ You may obtain a copy of the License at
-  ~
-  ~ http://www.apache.org/licenses/LICENSE-2.0
-  -->
-
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-		 xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-		 xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
-
-	<modelVersion>4.0.0</modelVersion>
-	<groupId>org.chenillekit</groupId>
-	<artifactId>chenillekit-image</artifactId>
-	<packaging>jar</packaging>
-
-	<parent>
-		<groupId>org.chenillekit</groupId>
-		<artifactId>chenillekit-project</artifactId>
-		<version>1.3.4-alpha-1</version>
-	</parent>
-
-	<name>Chenillekit Image</name>
-	<description>
-		A collection of sophisticated Ajax-enabled components based on the prototype/script.aculo.us JavaScript library
-		and many more ...
-	</description>
-
-	<scm>
-		<connection>scm:svn:http://svn.codehaus.org/chenillekit/trunk/${project.artifactId}/</connection>
-		<developerConnection>scm:svn:https://svn.codehaus.org/chenillekit/trunk/${project.artifactId}/</developerConnection>
-		<url>http://svn.codehaus.org/chenillekit/trunk/${project.artifactId}/</url>
-	</scm>
-
-	<!--
-	++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
-	+ dependencies for this module
-	++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
-	-->
-	<dependencies>
-
-		<dependency>
-			<groupId>org.apache.tapestry</groupId>
-			<artifactId>tapestry-test</artifactId>
-			<scope>test</scope>
-		</dependency>
-
-		<dependency>
-			<groupId>org.easymock</groupId>
-			<artifactId>easymock</artifactId>
-			<scope>test</scope>
-		</dependency>
-
-		<dependency>
-			<groupId>org.chenillekit</groupId>
-			<artifactId>chenillekit-test</artifactId>
-			<version>${project.version}</version>
-			<scope>test</scope>
-		</dependency>
-
-		<dependency>
-			<groupId>org.apache.tapestry</groupId>
-			<artifactId>tapestry-ioc</artifactId>
-		</dependency>
-
-		<dependency>
-			<groupId>com.google.code.kaptcha</groupId>
-			<artifactId>kaptcha</artifactId>
-			<version>2.3</version>
-			<scope>provided</scope>
-		</dependency>
-
-	</dependencies>
-
-	<build>
-		<plugins>
-
-			<!--
-			++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
-			+ pack the jar archive
-			++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
-			-->
-			<plugin>
-				<groupId>org.apache.maven.plugins</groupId>
-				<artifactId>maven-jar-plugin</artifactId>
-				<configuration>
-					<archive>
-						<compress>true</compress>
-						<index>true</index>
-						<manifestEntries>
-							<Tapestry-Module-Classes>
-								org.chenillekit.image.ChenilleKitImageModule
-							</Tapestry-Module-Classes>
-							<Implementation-Version>
-								${project.version}
-							</Implementation-Version>
-						</manifestEntries>
-					</archive>
-				</configuration>
-			</plugin>
-
-		</plugins>
-	</build>
-
-</project>
