--- chenillekit-image/src/main/java/org/chenillekit/image/ChenilleKitImageConstants.java
@@ -1,30 +1,0 @@
-/*
- * Apache License
- * Version 2.0, January 2004
- * http://www.apache.org/licenses/
- *
- * Copyright 2009 by chenillekit.org
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- */
-
-package org.chenillekit.image;
-
-/**
- * @version $Id$
- */
-public class ChenilleKitImageConstants
-{
-	public static String KAPATCHA_CONFIG_KEY = "chenillekit.kapatcha.properties";
-
-	private static ChenilleKitImageConstants ourInstance = new ChenilleKitImageConstants();
-
-    public static ChenilleKitImageConstants getInstance()
-    {
-        return ourInstance;
-    }
-}
