--- chenillekit-image/src/site/apt/index.apt
@@ -1,6 +1,0 @@
- ----
- Home
- ----
-
-Chenille Kit Image
-
