--- chenillekit-image/src/site/apt/kaptcha.apt
@@ -1,63 +1,0 @@
- ----
- ChenilleKit Captcha Service
- ----
-
-
-Configure the captcha service
-
-  To configure the captcha service, insert the contribution into your application module like the sample.
-  If you want to use the default values, leave it.
-
-+---+
-public static void contributeKaptchaProducer(MappedConfiguration<String, Properties> configuration)
-{
-	Properties properties = new Properties();
-
-	properties.setProperty("kaptcha.border", "no"); // Border around kaptcha. Legal values are yes or no.
-	...
-
-	configuration.add(ChenilleKitImageConstants.KAPATCHA_CONFIG_KEY, properties);
-}
-+---+
-
-Default captcha service configuration
-
-*----------------------------------*-------------------------------------------------------*---------------------------------------------------------------------------------------*
-| key                              | default value                                         | Description                                                                           |
-*----------------------------------+-------------------------------------------------------+---------------------------------------------------------------------------------------+
-| kaptcha.border                   | yes                                                   | Border around kaptcha. Legal values are yes or no.                                    |
-*----------------------------------+-------------------------------------------------------+---------------------------------------------------------------------------------------+
-| kaptcha.border.color             | black                                                 | Color of the border. Legal values are r,g,b (and optional alpha) or white,black,blue. |
-*----------------------------------+-------------------------------------------------------+---------------------------------------------------------------------------------------+
-| kaptcha.border.thickness         | 1                                                     | Thickness of the border around kaptcha. Legal values are > 0.                         |
-*----------------------------------+-------------------------------------------------------+---------------------------------------------------------------------------------------+
-| kaptcha.image.width              | 200                                                   | Width in pixels of the kaptcha image.                                                 |
-*----------------------------------+-------------------------------------------------------+---------------------------------------------------------------------------------------+
-| kaptcha.image.height             | 50                                                    | Height in pixels of the kaptcha image.                                                |
-*----------------------------------+-------------------------------------------------------+---------------------------------------------------------------------------------------+
-| kaptcha.producer.impl            | com.google.code.kaptcha.impl.DefaultKaptcha           | The image producer.                                                                   |
-*----------------------------------+-------------------------------------------------------+---------------------------------------------------------------------------------------+
-| kaptcha.textproducer.impl        | com.google.code.kaptcha.text.impl.DefaultTextCreator  | The text producer.                                                                    |
-*----------------------------------+-------------------------------------------------------+---------------------------------------------------------------------------------------+
-| kaptcha.textproducer.char.string | abcde2345678gfynmnpwx                                 | The text producer.                                                                    |
-*----------------------------------+-------------------------------------------------------+---------------------------------------------------------------------------------------+
-| kaptcha.textproducer.char.length | 5                                                     | The number of characters to display.                                                  |
-*----------------------------------+-------------------------------------------------------+---------------------------------------------------------------------------------------+
-| kaptcha.textproducer.font.names  | Arial, Courier                                        | A list of comma separated font names.                                                 |
-*----------------------------------+-------------------------------------------------------+---------------------------------------------------------------------------------------+
-| kaptcha.textproducer.font.size   | 40                                                    | The size of the font to use.                                                          |
-*----------------------------------+-------------------------------------------------------+---------------------------------------------------------------------------------------+
-| kaptcha.textproducer.font.color  | black                                                 | The color to use for the font. Legal values are r,g,b.                                |
-*----------------------------------+-------------------------------------------------------+---------------------------------------------------------------------------------------+
-| kaptcha.noise.impl               | com.google.code.kaptcha.impl.DefaultNoise             | The noise producer.                                                                   |
-*----------------------------------+-------------------------------------------------------+---------------------------------------------------------------------------------------+
-| kaptcha.noise.color              | black                                                 | The noise color. Legal values are r,g,b.                                              |
-*----------------------------------+-------------------------------------------------------+---------------------------------------------------------------------------------------+
-| kaptcha.obscurificator.impl      | com.google.code.kaptcha.impl.WaterRipple              | The obscurificator implementation.                                                    |
-*----------------------------------+-------------------------------------------------------+---------------------------------------------------------------------------------------+
-| kaptcha.background.impl          | com.google.code.kaptcha.impl.DefaultBackground        | The background implementation                                                         |
-*----------------------------------+-------------------------------------------------------+---------------------------------------------------------------------------------------+
-| kaptcha.background.clear.to      | white                                                 | Ending background color. Legal values are r,g,b                                       |
-*----------------------------------+-------------------------------------------------------+---------------------------------------------------------------------------------------+
-| kaptcha.word.impl                | com.google.code.kaptcha.text.impl.DefaultWordRenderer | The word renderer implementation                                                      |
-*----------------------------------+-------------------------------------------------------+---------------------------------------------------------------------------------------+
