--- chenillekit-image/src/site/site.xml
@@ -1,51 +1,0 @@
-<?xml version="1.0" encoding="ISO-8859-1"?>
-<!--
-  ~ Apache License
-  ~ Version 2.0, January 2004
-  ~ http://www.apache.org/licenses/
-  ~
-  ~ Copyright 2008 by chenillekit.org
-  ~
-  ~ Licensed under the Apache License, Version 2.0 (the "License");
-  ~ you may not use this file except in compliance with the License.
-  ~ You may obtain a copy of the License at
-  ~
-  ~ http://www.apache.org/licenses/LICENSE-2.0
-  -->
-
-<project name="ChenilleKit Image">
-
-    <bannerLeft>
-        <name>ChenilleKit Image</name>
-        <href>index.html</href>
-        <src>images/chenillekit-logo.png</src>
-    </bannerLeft>
-
-	<skin>
-		<groupId>org.chenillekit</groupId>
-		<artifactId>maven-skin</artifactId>
-		<version>1.0</version>
-	</skin>
-
-    <publishDate format="yyyy-MM-dd HH:mm:SS"/>
-
-    <body>
-
-        <links>
-            <item name="Project Home" href="../index.html"/>
-            <item name="Tapestry" href="http://tapestry.apache.org/"/>
-        </links>
-
-        <menu>
-            <item name="Project Home" href="../index.html"/>
-            <item name="Module Home" href="./index.html"/>
-        </menu>
-
-		<menu name="Services">
-			<item name="Captcha" href="kaptcha.html"/>
-		</menu>
-
-		<menu ref="reports"/>
-
-    </body>
-</project>
