--- chenillekit-image/src/test/java/org/chenillekit/image/ChenilleKitImageTestModule.java
@@ -1,25 +1,0 @@
-/*
- * Apache License
- * Version 2.0, January 2004
- * http://www.apache.org/licenses/
- *
- * Copyright 2008 by chenillekit.org
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- */
-
-package org.chenillekit.image;
-
-import org.apache.tapestry5.ioc.annotations.SubModule;
-
-/**
- * @version $Id$
- */
-@SubModule(value = {ChenilleKitImageModule.class})
-public class ChenilleKitImageTestModule
-{
-}
