--- chenillekit-image/src/test/resources/log4j.xml
@@ -1,39 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<!--
-  ~ Apache License
-  ~ Version 2.0, January 2004
-  ~ http://www.apache.org/licenses/
-  ~
-  ~ Copyright 2008 by chenillekit.org
-  ~
-  ~ Licensed under the Apache License, Version 2.0 (the "License");
-  ~ you may not use this file except in compliance with the License.
-  ~ You may obtain a copy of the License at
-  ~
-  ~ http://www.apache.org/licenses/LICENSE-2.0
-  -->
-
-<!DOCTYPE log4j:configuration SYSTEM "log4j.dtd">
-
-<log4j:configuration xmlns:log4j="http://jakarta.apache.org/log4j/">
-
-    <appender name="console" class="org.apache.log4j.ConsoleAppender">
-        <param name="Target" value="System.out"/>
-        <layout class="org.apache.log4j.PatternLayout">
-            <param name="ConversionPattern" value="%-5p %c{1} - %m%n"/>
-        </layout>
-    </appender>
-
-    <category name="org.chenillekit.template.ChenilleKitTemplateModule.VelocityService">
-        <priority value="trace"/>
-    </category>
-
-    <category name="org.chenillekit.template.ChenilleKitTemplateModule.FreeMarkerService">
-        <priority value="trace"/>
-    </category>
-
-    <root>
-        <priority value="warn"/>
-        <appender-ref ref="console"/>
-    </root>
-</log4j:configuration>
