+++ chenillekit-tapestry/pom.xml
@@ -26,7 +26,7 @@
 	<parent>
 		<groupId>org.chenillekit</groupId>
 		<artifactId>chenillekit-project</artifactId>
-		<version>1.3.4-alpha-1</version>
+		<version>1.3.4-alpha-2</version>
 	</parent>
 
 	<name>Chenillekit Tapestry</name>
@@ -59,12 +59,6 @@
 		<dependency>
 			<groupId>${project.groupId}</groupId>
 			<artifactId>chenillekit-core</artifactId>
-			<version>${project.parent.version}</version>
-		</dependency>
-
-		<dependency>
-			<groupId>${project.groupId}</groupId>
-			<artifactId>chenillekit-image</artifactId>
 			<version>${project.parent.version}</version>
 		</dependency>
