+++ pom.xml
@@ -20,7 +20,7 @@
 
 	<properties>
 		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-		<tapestry-release-version>5.3.5</tapestry-release-version>
+		<tapestry-release-version>5.3.6</tapestry-release-version>
 		<update-release-info>false</update-release-info>
 		<cobertura-plugin-version>2.3</cobertura-plugin-version>
 	</properties>
@@ -29,7 +29,7 @@
 	<groupId>org.chenillekit</groupId>
 	<artifactId>chenillekit-project</artifactId>
 	<packaging>pom</packaging>
-	<version>1.3.4-alpha-1</version>
+	<version>1.3.4-alpha-2</version>
 	<name>ChenilleKit Project</name>
 	<description>
 		A collection of sophisticated Ajax-enabled components based on the prototype/script.aculo.us JavaScript library
@@ -211,7 +211,6 @@
 		<module>chenillekit-demo</module>
 		<module>chenillekit-reports</module>
 		<module>chenillekit-template</module>
-		<module>chenillekit-image</module>
 		<module>chenillekit-google</module>
 		<module>chenillekit-scripting</module>
 		<module>chenillekit-mail</module>
